Add Asana panel
This commit disables the twitter panels, and replaces them with a simple
Asana panel. This adds three configuration settings for Asana, the
Access Token, a workspace name and the status of tasks you'd like to
display (defaults to today). In order to us Asana, you will need to
configure the Access Token in the Asana dashboard.