Get first two pages of Asana tasks
Asana only returns 50 tasks per page, and the tasks are in descending.
This commit make sure that we are getting up to 100 tasks, so that we
can filter as many as possible; for instance if we have 75 tasks, and
only the last 10 have the status 'today', retrieving the first page
of tasks will not display any tasks on the dashboard.
Additionally, changes the dashboard format as well, moving some of the
panels around.